+++ backend/tests/baserow/contrib/database/trash/test_database_trash_types.py
@@ -618,7 +618,7 @@ def test_restoring_a_trashed_link_field_restores_the_opposing_field_also(
         name="Customer",
         link_row_table=customers_table,
     )
-    TrashHandler.trash(user, database.group, database, link_field_1)
+    FieldHandler().delete_field(user, link_field_1)
 
     assert LinkRowField.trash.count() == 2
 
@@ -1024,18 +1024,21 @@ def test_a_restored_field_will_have_its_name_changed_to_ensure_it_is_unique(
     row_handler = RowHandler()
 
     # Create a primary field and some example data for the customers table.
-    customers_primary_field = field_handler.create_field(
-        user=user, table=customers_table, type_name="text", name="Name", primary=True
+    customers_field = field_handler.create_field(
+        user=user,
+        table=customers_table,
+        type_name="text",
+        name="Name",
     )
     row_handler.create_row(
         user=user,
         table=customers_table,
-        values={f"field_{customers_primary_field.id}": "John"},
+        values={f"field_{customers_field.id}": "John"},
     )
     row_handler.create_row(
         user=user,
         table=customers_table,
-        values={f"field_{customers_primary_field.id}": "Jane"},
+        values={f"field_{customers_field.id}": "Jane"},
     )
 
     link_field_1 = field_handler.create_field(
@@ -1045,8 +1048,8 @@ def test_a_restored_field_will_have_its_name_changed_to_ensure_it_is_unique(
         name="Customer",
         link_row_table=customers_table,
     )
-    TrashHandler.trash(user, database.group, database, link_field_1)
-    TrashHandler.trash(user, database.group, database, customers_primary_field)
+    field_handler.delete_field(user, link_field_1)
+    field_handler.delete_field(user, customers_field)
 
     assert LinkRowField.trash.count() == 2
 
@@ -1076,13 +1079,13 @@ def test_a_restored_field_will_have_its_name_changed_to_ensure_it_is_unique(
     assert link_field_2.link_row_related_field.name == "Table"
     assert link_field_1.link_row_related_field.name == "Table (Restored)"
 
-    TrashHandler.restore_item(user, "field", customers_primary_field.id)
-    customers_primary_field.refresh_from_db()
+    TrashHandler.restore_item(user, "field", customers_field.id)
+    customers_field.refresh_from_db()
 
     assert TextField.objects.count() == 3
     assert clashing_field.name == "Name"
     assert another_clashing_field.name == "Name (Restored)"
-    assert customers_primary_field.name == "Name (Restored) 2"
+    assert customers_field.name == "Name (Restored) 2"
 
     # Check that a normal trash and restore when there aren't any naming conflicts will
     # return the old names.
@@ -1128,9 +1131,7 @@ def test_perm_delete_related_link_row_field(data_fixture):
         link_row_table=customers_table,
     )
 
-    TrashHandler.trash(
-        user, database.group, database, link_field_1.link_row_related_field
-    )
+    field_handler.delete_field(user, link_field_1.link_row_related_field)
     assert TrashEntry.objects.count() == 1
     link_field_1.refresh_from_db()
     assert link_field_1.trashed